Fixed some comments

+++ Core/Source/Core/Rendering/LinearGradientBrush.cpp
@@ -106,7 +106,7 @@ namespace Core
 
 		HRESULT hr = rt->CreateGradientStopCollection(
 			gradientStops.data(),
-			gradientStops.size(),
+			(UINT32)gradientStops.size(),
 			&this->collection
 		);
 		if (FAILED(hr))
